Decomposing CharacterShooting into PlayerShooting and EnemyShooting
