+++ Assets/Prefabs/Enemies/EnemyType01.prefab
@@ -191,9 +191,9 @@ GameObject:
   - component: {fileID: 7661927029065131713}
   - component: {fileID: 2229548469702594169}
   - component: {fileID: -7393698408962043864}
-  - component: {fileID: 1131119205904769798}
   - component: {fileID: 5537784887623331358}
   - component: {fileID: 1735647611055947067}
+  - component: {fileID: 5771411646608422974}
   m_Layer: 0
   m_Name: EnemyType01
   m_TagString: Untagged
@@ -249,21 +249,6 @@ MonoBehaviour:
   <CurrentSpeed>k__BackingField: 0
   enemyStats: {fileID: 11400000, guid: 218ed6bce7588421381b29a87eabad23, type: 2}
   <CurrentEnemyState>k__BackingField: 0
---- !u!114 &1131119205904769798
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7661927029065131717}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3853fa1a5c201432cbf2c4bc8c337e4b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  minePrefab: {fileID: 7394722670251870173, guid: c87cd4df3d27240aa986a9cdcc999485, type: 3}
-  <CurrentWeapon>k__BackingField: {fileID: 3491157471432284435}
-  <AimingColor>k__BackingField: {r: 1, g: 0.023584902, b: 0.023584902, a: 0}
 --- !u!114 &5537784887623331358
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -301,6 +286,20 @@ NavMeshAgent:
   m_BaseOffset: 0.63
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
+--- !u!114 &5771411646608422974
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7661927029065131717}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 88966870957224ff3b4275141f6cba83, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  <CurrentWeapon>k__BackingField: {fileID: 6378564384715130769, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
+  <AimColor>k__BackingField: {r: 0, g: 0, b: 0, a: 0}
 --- !u!1001 &7157669485824485709
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -445,17 +444,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
---- !u!114 &3491157471432284435 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 6378564384715130769, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
-  m_PrefabInstance: {fileID: 7563271907803094658}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1359ae5aa984b4e1ab49f61467af9582, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &4760083955223552763 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 3096528867371367545, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
