+++ Assets/Prefabs/Player.prefab
@@ -211,8 +211,8 @@ GameObject:
   - component: {fileID: 5901563002306245278}
   - component: {fileID: 5901563002306245265}
   - component: {fileID: 5901563002306245264}
-  - component: {fileID: 7709062691085363798}
   - component: {fileID: 2344139235474058412}
+  - component: {fileID: 3523588330990207111}
   m_Layer: 6
   m_Name: Player
   m_TagString: Player
@@ -347,21 +347,6 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
---- !u!114 &7709062691085363798
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5901563002306245253}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3853fa1a5c201432cbf2c4bc8c337e4b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  minePrefab: {fileID: 7394722670251870173, guid: c87cd4df3d27240aa986a9cdcc999485, type: 3}
-  <CurrentWeapon>k__BackingField: {fileID: 3399491570507002170}
-  <AimingColor>k__BackingField: {r: 0.03208439, g: 0.19157214, b: 0.9716981, a: 0}
 --- !u!114 &2344139235474058412
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -377,6 +362,21 @@ MonoBehaviour:
   renderers:
   - {fileID: 5901563001997911701}
   - {fileID: 5901563000847378393}
+--- !u!114 &3523588330990207111
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5901563002306245253}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f244e061de9774b608e6a6ab3e2dff95, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  <CurrentWeapon>k__BackingField: {fileID: 6378564384715130769, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
+  <AimColor>k__BackingField: {r: 0, g: 0, b: 0, a: 0}
+  minePrefab: {fileID: 7394722670251870173, guid: c87cd4df3d27240aa986a9cdcc999485, type: 3}
 --- !u!1001 &8622243815050606251
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -446,17 +446,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
---- !u!114 &3399491570507002170 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 6378564384715130769, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
-  m_PrefabInstance: {fileID: 8622243815050606251}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1359ae5aa984b4e1ab49f61467af9582, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &6724236569161485010 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 3096528867371367545, guid: 5a5bb90417dae4efd87d1426a0dd1449, type: 3}
