+++ Assets/Prefabs/Weapon/WeaponType01.prefab
@@ -28,7 +28,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 33142460980013606}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalRotation: {x: 0, y: 0.3074675, z: 0, w: 0.9515586}
   m_LocalPosition: {x: 55.751, y: 15.3, z: 303.886}
   m_LocalScale: {x: 0.2, y: 0.2, z: 0.9}
   m_ConstrainProportionsScale: 0
