+++ Assets/Scripts/CharacterShooting.cs
@@ -1,8 +1,6 @@
-using System;
 using UnityEngine;
 using DmitryAdventure.Armament;
 using UnityEngine.InputSystem;
-using Random = UnityEngine.Random;
 
 // ReSharper disable once CheckNamespace
 namespace DmitryAdventure.Characters
@@ -10,53 +8,33 @@ namespace DmitryAdventure.Characters
     /// <summary>
     /// Represents character's shooting functionality.
     /// </summary>
-    public class CharacterShooting : MonoBehaviour
+    public abstract class CharacterShooting : MonoBehaviour
     {
         #region Сonstants, variables & properties
-
-        [SerializeField] private Mine minePrefab;
-
-        [field:SerializeField, Tooltip("Character's chosen weapon")] 
-        private Weapon CurrentWeapon { get; set; }
-
+        
+        [field:SerializeField, ReadonlyField, Tooltip("Character's chosen weapon")] 
+        protected Weapon CurrentWeapon { get; set; }
+        
+        protected Vector3 AimPoint;
+        
         [field: SerializeField, Tooltip("Sight beam color (for Gismo)")]
-        private Color AimingColor { get; set; }
+        private Color AimColor { get; set; }
+        
+        protected Character Character;
 
-        private Vector3 _aimingPoint;
-        
-        private Camera _camera;
-        private Character _character;
-        private CharacterInventory _characterInventory;
-        
-        private PlayerInput _playerInput;
-        private InputAction _characterFireAction;
-        private InputAction _characterMineAction;
         #endregion
         
         #region Monobehavior methods
 
-        private void Awake()
+        protected virtual void Awake()
         {
-            _camera = Camera.main;
-            _character = gameObject.GetComponent<Character>();
-
-            if (_character.CharacterType != CharacterType.Player) return;
-            
-            _characterInventory = gameObject.GetComponent<CharacterInventory>();
-            _playerInput = gameObject.GetComponent<PlayerInput>();
+            CurrentWeapon = GetComponentInChildren<Weapon>();
+            Character = gameObject.GetComponent<Character>();
         }
         
-        private void Start()
+        protected virtual void Start()
         {
-            _aimingPoint = Vector3.zero;
-
-            if (_character.CharacterType != CharacterType.Player) return;
-            
-            _characterFireAction = _playerInput.actions["Fire"];
-            _characterFireAction.performed += ShootWeapon;
-            
-            _characterMineAction = _playerInput.actions["Mine"];
-            _characterMineAction.performed += MineActionOnPerformed;
+            AimPoint = Vector3.zero;
         }
 
         private void Update()
@@ -66,112 +44,25 @@ namespace DmitryAdventure.Characters
 
         private void OnDrawGizmos()
         {
-            if (_aimingPoint == Vector3.zero) return;
+            if (AimPoint == Vector3.zero) return;
             
-            Gizmos.color = AimingColor;
-            Gizmos.DrawLine(CurrentWeapon.ShotPoint.position, _aimingPoint!);
+            Gizmos.color = AimColor;
+            Gizmos.DrawLine(CurrentWeapon.ShotPoint.position, AimPoint!);
         }
-
-        private void OnDestroy()
-        {
-            if (_character.CharacterType != CharacterType.Player) return;
-            
-            _characterFireAction.performed -= ShootWeapon;
-            _characterMineAction.performed -= MineActionOnPerformed;
-        }
-
-        #endregion
-
-      
-
-
-        #region Event handlers
-
-        /// <summary>
-        /// Handler for selecting a mine from inventory.
-        /// </summary>
-        /// <param name="context">CallbackContext for more info.</param>
-        private void MineActionOnPerformed(InputAction.CallbackContext context)
-        {
-            if(_characterInventory == null) return;
-            
-            var popMine = _characterInventory.PopFromInventory(GameData.MineKey);
-            if (popMine == null) return;
-
-            Instantiate(minePrefab, new Vector3(_aimingPoint.x,_aimingPoint.y + 0.2f,_aimingPoint.z), Quaternion.identity);
-        }
-        
         #endregion
 
         #region Other methods
 
-        /// <summary>
-        /// Aiming character.
-        /// </summary>
-        private void OnTakeAim()
-        {
-            switch (_character.CharacterType)
-            {
-                case CharacterType.Player:
-                {
-                    var camTransform = _camera.transform;
-            
-                    if (!Physics.Raycast(camTransform.position, camTransform.forward, out var hit)) 
-                        return;
+        protected abstract void OnTakeAim();
 
-                    var targetDistance = Vector3.Distance( CurrentWeapon.ShotPoint.position, hit.point);
-                    if (targetDistance < CurrentWeapon.weaponStats.ShotRange)
-                    {
-                        _aimingPoint = hit.point;
-                    }
-                    else
-                    {
-                        _aimingPoint = camTransform.position + camTransform.forward * CurrentWeapon.weaponStats.ShotRange;
-                    }
-                    break;
-                }
-                case CharacterType.EnemyType01:
-                    
-                    var enemy = _character as Enemy;
-                    if (enemy == null) return;
-                    
-                    if (enemy.CurrentEnemyState == EnemyState.Attack)
-                    {
-                        var eObject = enemy.gameObject;
-                        var ePosition = eObject.transform.position;
-                        var eForward = eObject.transform.forward;
-                        var layerMask = GameData.PlayerLayerMask;
-                        
-                        if (Physics.Raycast(
-                                ePosition,
-                                eForward,
-                                out var playerHit,
-                                enemy.enemyStats.AttentionRadius,
-                                layerMask))
-                        {
-                            var bounds = playerHit.collider.bounds;
-                            
-                            _aimingPoint = new Vector3(
-                                Random.Range(bounds.min.x, bounds.max.x),
-                                Random.Range(bounds.min.y, bounds.max.y),
-                                Random.Range(bounds.min.z, bounds.max.z)
-                            );
-                            ShootWeapon();
-                        }
-                    }
-
-                    break;
-            }
-        }
-        
         /// <summary>
         /// Fires a shot from a weapon.
         /// </summary>
         /// <param name="context">CallbackContext for more info.</param>
         /// <remarks>For player.</remarks>>
-        private void ShootWeapon(InputAction.CallbackContext context)
+        protected void ShootWeapon(InputAction.CallbackContext context)
         {
-            CurrentWeapon.Fire(_aimingPoint);
+            CurrentWeapon.Fire(AimPoint);
         }
         
         /// <summary>
@@ -180,10 +71,9 @@ namespace DmitryAdventure.Characters
         /// <remarks>For NPC.</remarks>>
         protected void ShootWeapon()
         {
-            CurrentWeapon.Fire(_aimingPoint);
+            CurrentWeapon.Fire(AimPoint);
         }
 
         #endregion
-  
     }
 }
