+++ Assets/Scripts/EnemyShooting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 88966870957224ff3b4275141f6cba83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
