+++ Assets/Scripts/PlayerShooting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f244e061de9774b608e6a6ab3e2dff95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
